pyproject.toml
[tool.poetry]
name = "productapi"
version = "0.1.0"
description = "Тестовое задание для Строй-Отделка"
authors = ["Starobogatov Alexey <Starobogatov.a@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.1"}
SQLAlchemy = "^2.0.32"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
bcrypt = "^4.2.0"
gunicorn = "^23.0.0"
PyJWT = "^2.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
